I've made the initial User interface skeleton
With this push, I defined the project to be a server application.  More on how to connect the user interface with the server running the model will be added further.  The user interface I've made defines the layout and some basic functions to connect pages, the design is just conceptual and it will be changed to please the user visual experience in the end.